--- a/packages/core/integration-tests/test/integration/babel-node-modules-jsx/.eslintrc.json
+++ b/packages/core/integration-tests/test/integration/babel-node-modules-jsx/.eslintrc.json
@@ -1,0 +1,6 @@
+{
+  "extends": "../.eslintrc.json",
+  "parserOptions": {
+    "sourceType": "module"
+  }
+}