--- a/packages/core/integration-tests/test/integration/babel-node-modules-jsx/index.js
+++ b/packages/core/integration-tests/test/integration/babel-node-modules-jsx/index.js
@@ -1,0 +1,4 @@
+import Foo from 'foo';
+
+export {Foo};
+export class Bar {}