--- a/packages/core/integration-tests/test/integration/babel-node-modules-jsx/package.json
+++ b/packages/core/integration-tests/test/integration/babel-node-modules-jsx/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "parcel-test-browser-browserslist",
+  "private": true,
+  "browserslist": ["last 2 Chrome versions", "IE >= 11"]
+}