--- a/packages/core/integration-tests/test/transpilation.js
+++ b/packages/core/integration-tests/test/transpilation.js
@@ -67,6 +67,7 @@ describe('transpilation', function () {
   });
 
   it('should not support JSX in node_modules', async function () {
+    // $FlowFixMe
     assert.rejects(
       bundle(
         path.join(__dirname, '/integration/babel-node-modules-jsx/index.js'),