Fixes #1655. Fixes #6749.

This enables transpiling node_modules using SWC's version of preset-env. Therefore, only standard JS syntax is allowed (no JSX, TypeScript, etc.). However, this is important as more modules on npm start publishing modern syntax rather than ES5. And now that we are using SWC and not Babel, the performance hit should be minimal.

Needs testing on real world projects (please help!), and perf testing as well.