@devongovett 2022-01-03
Some benchmark data for comparison. All done with the Atlaskit Editor benchmark in the parcel-benchmark-action repo, with `--no-optimize`.

Before (v2): 5.24s
After: 9.46s (browserslist: ">= 0.25%")
After: 6.04s (browserslist: "supports es6-module")
After: 5.44s (browserslist: "last 2 Chrome versions")

With `>= 0.25%`, the build becomes ~2x slower because we compile 2x more code to produce the `nomodule` version. But with the `supports es6-module` query, the build is only a small amount slower.